Face_Recognition/requirements.txt
face-recognition==1.3.0
opencv-python==4.8.0.74
Flask==2.3.3
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.0 